rearrange the project folder/file structure and rename the component
