+++ src/App.js
@@ -3,16 +3,15 @@ import 'semantic-ui-css/semantic.min.css';
 import 'react-toastify/dist/ReactToastify.css';
 import 'cropperjs/dist/cropper.css';
 import { Grid, Segment } from 'semantic-ui-react';
-import NavHeader from './Components/Nav/NavHeader';
-import NavHeaderMobile from './Components/Nav/NavHeaderMobile';
-import CharacterToDo from './Components/Home/CharacterToDo';
-import CharacterToDoRow from './Components/Home/CharacterToDoRow';
-import CharacterToDoRowMobile from './Components/Home/CharacterToDoRowMobile';
-import Login from './Components/Home/Login';
-import Register from './Components/Home/Register';
+import NavHeader from './pages/navHeader/web/NavHeader';
+import NavHeaderMobile from './pages/navHeader/mobile/NavHeaderMobile';
+import CharacterToDoRow from './pages/userHomework/index';
+import CharacterToDoRowMobile from './pages/userHomeworkMobile/index';
+import Login from './pages/login/index';
+import Register from './pages/register/index';
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import { createMedia } from '@artsy/fresnel';
-import Dashboard from './Components/Dashboard/Dashboard';
+import Dashboard from './pages/dashboard/index';
 
 const AppMedia = createMedia({
   breakpoints: { zero: 0, mobile: 549, tablet: 850, computer: 1080 },
