--- src/Components/Todo/TodoComponent.css
@@ -1,11 +1,0 @@
-.iconButton {
-  cursor: pointer;
-}
-
-.thisItemNotChanged {
-  border-top: 0.05rem inset ivory;
-}
-
-.thisItemChanged {
-  border: 0.05rem inset red;
-}
