+++ src/pages/components/characterRelated/CharacterAvatar.css
