+++ src/pages/components/util/ViewDataUtil.js
