+++ src/pages/components/util/uploadPicToCloudinary.js
@@ -6,7 +6,6 @@ const uploadPic = async (media) => {
     form.append("file", media);
     form.append("upload_preset", "Loado_Image");
     form.append("cloud_name", "biglol");
-    console.log("URL IS ", process.env.REACT_APP_CLOUDINARY_URL);
     const res = await axios.post(process.env.REACT_APP_CLOUDINARY_URL, form);
     return res.data.url;
   } catch (error) {
