+++ src/pages/register/index.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import {
   Button,
   Form,
@@ -11,16 +11,17 @@ import {
 import axios from 'axios';
 import cookie from 'js-cookie';
 import { Link, useHistory } from 'react-router-dom';
-import { backendUrl, axiosConfig } from '../Utils/ConstVar';
+import { backendUrl, axiosConfig } from '../components/util/ConstVar';
 
-import './Login.css';
+import '../login/index.css';
 
-function Login() {
+function Register() {
   const history = useHistory();
-  const [loginMessage, setLoginMessage] = useState('로그인이 필요합니다');
+  const [loginMessage, setLoginMessage] = useState('회원가입');
   const [user, setUser] = useState({
-    id: cookie.get('loginId') ? cookie.get('loginId') : '',
+    id: '',
     password: '',
+    name: '',
   });
   const handleChange = (e) => {
     const { name, value } = e.target;
@@ -31,10 +32,11 @@ function Login() {
     e.preventDefault();
     axios
       .post(
-        `${backendUrl}/loado/api/users/login`,
+        `${backendUrl}/loado/api/users/register`,
         {
           userId: user.id,
           password: user.password,
+          name: user.name,
         },
         axiosConfig
       )
@@ -51,17 +53,9 @@ function Login() {
         }
       })
       .catch((err) => {
-        // console.log(err.request);
-        // console.log(err.response);
-        // console.log(err.message);
         setLoginMessage(err.response.data.error);
       });
   };
-
-  useEffect(() => {
-    let loginCookie = cookie.get('loadoUserToken');
-    loginCookie && history.push('/userhomework');
-  }, []);
 
   return (
     <div id='divLoginPage'>
@@ -96,17 +90,27 @@ function Login() {
                 name='password'
                 onChange={handleChange}
               />
+              <Form.Input
+                fluid
+                icon='quote left'
+                iconPosition='left'
+                placeholder='이름'
+                name='name'
+                onChange={handleChange}
+                value={user.name}
+              />
 
               <Button color='teal' fluid size='large' type='submit'>
-                로그인
+                가입
               </Button>
             </Segment>
           </Form>
           <Message>
-            계정이 없으신가요?{' '}
-            <a href='#' onClick={() => history.push('/register')}>
-              가입
-            </a>
+            비밀번호를 잊어버리시면 복구가 불가능합니다
+            <br />
+            [비밀번호는 암호화되어 저장됩니다]
+            <br />
+            [예시: .....APvs6TgQOG5d8byxYdLJOD0O6Am.....]
           </Message>
         </Grid.Column>
       </Grid>
@@ -137,4 +141,4 @@ function Login() {
   );
 }
 
-export default Login;
+export default Register;
