+++ src/pages/userHomework/index.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
-import "./CharacterToDo.css";
-import AddCharacter from "../Utils/AddCharacter";
+import "./index.css";
+import AddCharacter from "../components/characterRelated/AddCharacter";
 
 import {
   Segment,
@@ -11,15 +11,15 @@ import {
   Loader,
   Image,
 } from "semantic-ui-react";
-import RestValue from "../GridItem/RestValue";
-import CharacterAvatar from "./CharacterAvatar";
-import PerIdNote from "./PerIdNote";
+import RestValue from "../components/userHomework/RestValue";
+import CharacterAvatar from "../components/characterRelated/CharacterAvatar";
+import PerIdNote from "../components/userHomework/PerIdNote";
 import {
   ChaosDunValue,
   GuardianDunValue,
   EponaValue,
   WeeklyGuardian,
-} from "../GridItem/DungeonAndEpona";
+} from "../components/userHomework/DungeonAndEpona";
 import {
   AbyssDun2,
   ArgosRaid,
@@ -27,14 +27,14 @@ import {
   BiakissRaid,
   KukseitnRaid,
   AbrelRaid,
-} from "../GridItem/AbyssAndRaid";
+} from "../components/userHomework/AbyssAndRaid";
 import axios from "axios";
 import cookie from "js-cookie";
 
 import { ToastContainer } from "react-toastify";
 import { useHistory } from "react-router-dom";
 
-import { backendUrl, axiosConfigAuth } from "../Utils/ConstVar";
+import { backendUrl, axiosConfigAuth } from "../components/util/ConstVar";
 import {
   viewDataMain,
   applyChangesUtil,
@@ -42,11 +42,11 @@ import {
   toastMessage,
   getUserCheckBoxConfiguration,
   changeUserCheckBoxConfiguration,
-} from "../Utils/ViewDataUtil";
-import AddAndChange from "../HomeSubComp/AddAndChange";
-import PaginationComp from "../HomeSubComp/PaginationComp";
-import SettingChange from "../HomeSubComp/SettingChange";
-import AlarmAndNote from "../HomeSubComp/AlarmAndNote";
+} from "../components/util/ViewDataUtil";
+import AddAndChange from "../components/userHomework/AddAndChange";
+import PaginationComp from "../components/userHomework/PaginationComp";
+import SettingChange from "../components/userHomework/SettingChange";
+import AlarmAndNote from "../components/userHomework/AlarmAndNote";
 
 function CharacterToDoRow({ limit, type }) {
   const todayDate = new Date(Date.now());
